requirements.txt
py-cord
pillow
asyncpraw